stream_filter/requirements.txt
idna==3.10
confluent-kafka==2.10.0
python-schema-registry-client[faust]
httpx
avro
requests
fastavro
cachetools
jsonschema
authlib
referencing
python-dotenv==1.0.0
intervaltree==3.1.0
mode-streaming==0.4.1
propcache==0.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
rocksdict==0.3.24
attrs==24.2.0
click==8.0.0
colorlog==6.9.0
croniter==2.0.7
faust-streaming==0.11.3
faust-streaming-rocksdb==0.9.3
frozenlist==1.5.0
six==1.17.0
sortedcontainers==2.4.0
terminaltables==3.1.10
typing_extensions==4.12.2
venusian==3.1.0
yarl==1.18.3
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
multidict==6.1.0
mypy-extensions==1.0.0
opentracing==1.3.0
packaging==23.2
aiohttp-cors==0.7.0
aiokafka==0.12.0
aiosignal==1.3.1
async-timeout==5.0.1
